+++ server.js
@@ -14,7 +14,7 @@ var app = express();
 
 var session = require('client-sessions');
 var getDB = require('./connect.js');
-var extra = require('./functions.js')
+
 // handlebars setup
 app.set('view engine', 'hbs');
 hbs.registerPartials(__dirname + '/views/partials');
@@ -35,8 +35,26 @@ app.use(session({
     activeDuration: 2 * 60 * 1000
 }));
 
+function login(email, password, callback) {
+    if (email.indexOf('@') > 0 && email.indexOf('.') > 0 && (email.indexOf('com') > 0 || email.indexOf('ca') > 0)) {
+        getDB.readFile(email, (err, user) => {
+            if(user === 'failed') {
+                callback(err, 'failed')
+            } else {
+                if (password === user.password) {
+                    callback(err, user)
+                } else {
+                    callback(err, 'failed')
+                }
+            }
+        }); 
+    } else {
+        callback('failed')
+    }
+}
+
 app.post('/login', function(req, res) {
-    extra.login(req.body.email, req.body.password, (err, user) => {
+    login(req.body.email, req.body.password, (err, user) => {
         if (user === 'failed') {
             res.render('login.hbs', {
                 error: 'Wrong email or password'
